Change Float to CGFloat where it relates to graphics logic

+++ LazyTransitions/ScrollableGestureHandler.swift
@@ -37,7 +37,7 @@ public class ScrollableGestureHandler: TransitionGestureHandler {
         handleTransitionGestureBegan(gesture)
     }
     
-    public func calculateTransitionProgressWithTranslation(_ translation: CGPoint, on view: UIView?) -> Float {
+    public func calculateTransitionProgressWithTranslation(_ translation: CGPoint, on view: UIView?) -> CGFloat {
         guard let view = view else { return 0 }
         
         let progress = TransitionProgressCalculator
